--- src/deluge/dsp/reverb/reverb_manager.cpp
@@ -1,1 +1,0 @@
-#include "reverb_manager.hpp"
