+++ src/deluge/dsp/reverb/valley/InterpDelay.hpp
@@ -10,6 +10,11 @@
 #include <cstddef>
 #include <cstdint>
 
+template<typename T>
+T linterp(T a, T b, T f) {
+    return a + f * (b - a);
+}
+
 template <typename T = float, size_t maxLength = 512>
 class InterpDelay {
 public:
