+++ src/deluge/gui/l10n/strings.h
@@ -706,6 +706,11 @@ enum class String : size_t {
 	STRING_FOR_PAD_SELECTION_OFF,
 	STRING_FOR_PAD_SELECTION_ON,
 
+	STRING_FOR_MODEL,
+	STRING_FOR_FREEVERB,
+	STRING_FOR_MUTABLE,
+	STRING_FOR_LEXICON_PLATE,
+
 	STRING_LAST
 };
