+++ src/deluge/gui/views/view.cpp
@@ -19,6 +19,7 @@
 #include "definitions_cxx.hpp"
 #include "deluge/model/settings/runtime_feature_settings.h"
 #include "dsp/reverb/freeverb/revmodel.hpp"
+#include "dsp/reverb/reverb.hpp"
 #include "extern.h"
 #include "gui/colour.h"
 #include "gui/context_menu/clear_song.h"
@@ -1262,8 +1263,8 @@ void View::pretendModKnobsUntouchedForAWhile() {
 
 void View::cycleThroughReverbPresets() {
 
-	int32_t currentRoomSize = AudioEngine::reverb.getroomsize() * 50;
-	int32_t currentDampening = AudioEngine::reverb.getdamp() * 50;
+	int32_t currentRoomSize = AudioEngine::reverb.get_room_size() * 50;
+	int32_t currentDampening = AudioEngine::reverb.get_damping() * 50;
 
 	// See which preset we're the closest to currently
 	int32_t lowestDifferentness = 1000;
@@ -1282,8 +1283,8 @@ void View::cycleThroughReverbPresets() {
 		newPreset = 0;
 	}
 
-	AudioEngine::reverb.setroomsize((float)presetReverbRoomSize[newPreset] / 50);
-	AudioEngine::reverb.setdamp((float)presetReverbDampening[newPreset] / 50);
+	AudioEngine::reverb.set_room_size((float)presetReverbRoomSize[newPreset] / 50);
+	AudioEngine::reverb.set_damping((float)presetReverbDampening[newPreset] / 50);
 
 	display->displayPopup(deluge::l10n::get(presetReverbNames[newPreset]));
 }
